added setMyUsername and fixed int id problem

+++ service/api/do-login.go
@@ -24,7 +24,7 @@ func (rt *_router) doLogin(w http.ResponseWriter, r *http.Request, ps httprouter
 
 	userID, err := rt.db.GetUser(username)
 	if err == nil {
-		response := map[string]int{"userID": userID}
+		response := map[string]uint64{"userID": userID}
 		w.Header().Set("Content-Type", "application/json")
 		json.NewEncoder(w).Encode(response)
 		return
@@ -37,7 +37,7 @@ func (rt *_router) doLogin(w http.ResponseWriter, r *http.Request, ps httprouter
 		return
 	}
 
-	response := map[string]int{"userID": newUserID}
+	response := map[string]uint64{"userID": newUserID}
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(response)
 }
