+++ service/database/delete-ban.go
@@ -1,7 +1,7 @@
 package database
 
 // GetName is an example that shows you how to query data
-func (db *appdbimpl) DeleteBan(IdUserToUnban int, IdUser int) error {
+func (db *appdbimpl) DeleteBan(IdUserToUnban uint64, IdUser uint64) error {
 	_, err := db.c.Exec(
 		"DELETE FROM Follows WHERE BannerId = ? AND BannedId = ?", IdUser, IdUserToUnban)
 	return err
