+++ service/database/delete-follow.go
@@ -1,7 +1,7 @@
 package database
 
 // GetName is an example that shows you how to query data
-func (db *appdbimpl) DeleteFollow(IdUserToNotFollow int, IdFollowingUser int) error {
+func (db *appdbimpl) DeleteFollow(IdUserToNotFollow uint64, IdFollowingUser uint64) error {
 	_, err := db.c.Exec(
 		"DELETE FROM Follows WHERE FollowerId = ? AND FollowedId = ?", IdFollowingUser, IdUserToNotFollow)
 	return err
