+++ service/database/delete-like.go
@@ -1,7 +1,7 @@
 package database
 
 // GetName is an example that shows you how to query data
-func (db *appdbimpl) DeleteLike(IdPhoto int, UserLikeId int) error {
+func (db *appdbimpl) DeleteLike(IdPhoto uint64, UserLikeId uint64) error {
 	_, err := db.c.Exec(
 		"DELETE FROM Likes WHERE PhotoId = ? AND UserId = ?", IdPhoto, UserLikeId)
 	return err
