+++ service/database/delete-photo.go
@@ -1,6 +1,6 @@
 package database
 
-func (db *appdbimpl) DeletePhoto(id int) error {
+func (db *appdbimpl) DeletePhoto(id uint64) error {
 	_, err := db.c.Exec("DELETE FROM Photos WHERE Id = ?", id)
 	return err
 }
